rmc/mongo-db/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
flet
